requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
asgiref==3.10.0
    # via django
cachetools==6.2.2
    # via google-auth
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
dj-database-url==3.0.1
    # via quizzer-ai (pyproject.toml)
django==5.2.8
    # via
    #   quizzer-ai (pyproject.toml)
    #   dj-database-url
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via google-generativeai
google-auth==2.43.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.1
    # via google-api-python-client
google-generativeai==0.8.5
    # via quizzer-ai (pyproject.toml)
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
gunicorn==23.0.0
    # via quizzer-ai (pyproject.toml)
h11==0.16.0
    # via uvicorn
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via requests
packaging==25.0
    # via gunicorn
pillow==12.0.0
    # via quizzer-ai (pyproject.toml)
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.11
    # via quizzer-ai (pyproject.toml)
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.4
    # via google-generativeai
pydantic-core==2.41.5
    # via pydantic
pyparsing==3.2.5
    # via httplib2
python-dotenv==1.2.1
    # via quizzer-ai (pyproject.toml)
requests==2.32.5
    # via google-api-core
rsa==4.9.1
    # via google-auth
sqlparse==0.5.3
    # via django
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via django
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via quizzer-ai (pyproject.toml)
whitenoise==6.11.0
    # via quizzer-ai (pyproject.toml)
